# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend App Ecommerce API",
        "description": "Documentación de la API del Ecommerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:3001",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Listar personas (Proveedores/Entidades)",
                "description": "Obtiene una lista paginada de personas. Permite buscar por nombres, apellidos o número de documento.",
                "operationId": "90c467ad608bc37bd7820a478b635ac2",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda (nombres, apellidos, número de documento).",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Juan"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10,
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personas obtenidas correctamente"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Registrar persona",
                "description": "Crea el registro de una persona, con los nuevos campos de contacto y ubigeo.",
                "operationId": "2f086ff8c55311b5938907cc196d5173",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_documento",
                                    "nombres"
                                ],
                                "properties": {
                                    "tipo_documento_identidad_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "numero_documento": {
                                        "type": "string",
                                        "example": "20123456789",
                                        "maxLength": 20
                                    },
                                    "nombres": {
                                        "type": "string",
                                        "example": "Comercializadora SAC",
                                        "maxLength": 150
                                    },
                                    "apellido_paterno": {
                                        "type": "string",
                                        "example": "Pérez",
                                        "nullable": true,
                                        "maxLength": 100
                                    },
                                    "apellido_materno": {
                                        "type": "string",
                                        "example": "García",
                                        "nullable": true,
                                        "maxLength": 100
                                    },
                                    "numero_celular": {
                                        "type": "string",
                                        "example": "999888777",
                                        "nullable": true,
                                        "maxLength": 20
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "contacto@comercial.com",
                                        "nullable": true,
                                        "maxLength": 150
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Las Palmeras 123",
                                        "nullable": true,
                                        "maxLength": 255
                                    },
                                    "departamento_id": {
                                        "type": "integer",
                                        "example": 15,
                                        "nullable": true
                                    },
                                    "provincia_id": {
                                        "type": "integer",
                                        "example": 1501,
                                        "nullable": true
                                    },
                                    "distrito_id": {
                                        "type": "integer",
                                        "example": 150101,
                                        "nullable": true
                                    },
                                    "estado": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona registrada correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/personas/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener detalle de persona",
                "description": "Obtiene la información completa de una persona por su ID.",
                "operationId": "a9ce473f12514e4a1a8da5706ef3a698",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos obtenidos correctamente"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualizar persona",
                "description": "Actualiza los datos personales y de ubicación de la persona.",
                "operationId": "01988886b1ad660e830db7a0c81f670a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_documento",
                                    "nombres"
                                ],
                                "properties": {
                                    "tipo_documento_identidad_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "numero_documento": {
                                        "type": "string",
                                        "example": "20123456789",
                                        "maxLength": 20
                                    },
                                    "nombres": {
                                        "type": "string",
                                        "example": "Comercializadora SAC",
                                        "maxLength": 150
                                    },
                                    "apellido_paterno": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 100
                                    },
                                    "apellido_materno": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 100
                                    },
                                    "numero_celular": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 20
                                    },
                                    "correo": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 150
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 255
                                    },
                                    "departamento_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "provincia_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "distrito_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "estado": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona actualizada correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Eliminar persona",
                "description": "Eliminación lógica (SoftDelete) de una persona.",
                "operationId": "fbf9123f9be359bc2878e23157d26269",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona eliminada correctamente"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT con el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}